+++ e2e-tests/ui-components/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @evervault/ui-components-e2e-tests
 
+## 1.0.7
+
+### Patch Changes
+
+- @evervault/ui-components@1.12.1
+
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [2eb3bf3]
+  - @evervault/ui-components@1.12.0
+
 ## 1.0.5
 
 ### Patch Changes