+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @evervault/ui-components
 
+## 1.12.1
+
+### Patch Changes
+
+- Updated dependencies [1f6edc4]
+  - @evervault/card-validator@1.1.0
+  - shared@1.0.7
+
+## 1.12.0
+
+### Minor Changes
+
+- 2eb3bf3: Automatically pad expiry month when a user enters a number betwen 2-9
+
 ## 1.11.0
 
 ### Minor Changes